move map outside of algo
